backend/requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.25.0
Pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
safetensors>=0.3.0